% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{igkv_mus_df}
\alias{igkv_mus_df}
\title{igkv_mus_df}
\format{A data frame with 76 rows and 2 variables:
\describe{
  \item{gene}{The gene name}
  \item{seq}{The corresponding sequence}
}}
\source{
IMGT
}
\usage{
igkv_mus_df
}
\description{
mouse germline IgK-V (light chain kappa V-gene segments). When multiple alleles were present, the
first one was included. These names and sequences can be changed by
customized by changing this dataframe. Additionally, repeating elements
can give certain germline gene elements a larger probability of being used
during repertoire evolution.
}
\keyword{datasets}

