% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{iglv_mus_df}
\alias{iglv_mus_df}
\title{iglv_mus_df}
\format{A data frame with 6 rows and 10 variables:
\describe{
  \item{gene}{The gene name}
  \item{seq}{The corresponding sequence}
}}
\source{
IMGT
}
\usage{
iglv_mus_df
}
\description{
mouse germline IgL-V (light chain lambda V-gene segments). When multiple alleles were present, the
first one was included. These names and sequences can be changed by
customized by changing this dataframe. Additionally, repeating elements
can give certain germline gene elements a larger probability of being used
during repertoire evolution.
}
\keyword{datasets}

